==> Scripts/Final_Scripts/cvnp_analyses-RETAIN BUT NOT IN THIS REPOSITORY/1_trajectoryresults.Rmd <==
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r echo=FALSE, include=FALSE}
library(RColorBrewer)
library(lattice)
library(tidyverse)
library(here)
library(tictoc)

library(ggplot2)
library(ggpubr)
library(sf)
library(terra)

library(tmap)
library(tmaptools)
library(RColorBrewer)
library(viridis)

terraOptions(memfrac=0.8, tempdir = here("Scratch"), progress=10)
```
#Introduction
In this script, I analyse the trajectories across indices and monthly/annual in CVNP
(1) Data input and crop/mask to CVNP
(2) Map making of trajectories
(3) Barplot of area results for each trajetory

#Step 1- data input, crop/mask
```{r}
##Data input
#CVNP shp
cvnp_shp<- st_read(here("Data", "Admin", "WDPA_WDOECM_Jan2024_Public_59_shp","WDPA_WDOECM_Jan2024_Public_59_shp_0", "WDPA_WDOECM_Jan2024_Public_59_shp-polygons.shp")) 

monthly_input_file_path <- here("Outputs", "TrendsResults", "results_rasters", "monthly")
monthly_file_list<- list.files(path = paste0(monthly_input_file_path, "/"), pattern = paste0("*","_monthly_swin11","*"), all.files = T, full.names = T)
monthly_file_list <- gtools::mixedsort(monthly_file_list)

monthly_rds_list<- lapply(monthly_file_list, rast)
names(monthly_rds_list[[1]])<- "evi_swin11_monthly"
names(monthly_rds_list[[2]])<- "ndvi_swin11_monthly"

annual_input_file_path <- here("Outputs", "TrendsResults", "results_rasters", "annual")
annual_file_list<- list.files(path = paste0(annual_input_file_path, "/"), pattern = paste0("*","_annual_swin11","*"), all.files = T, full.names = T)
annual_file_list <- gtools::mixedsort(annual_file_list)

annual_rds_list<- lapply(annual_file_list, rast)
names(annual_rds_list[[1]])<- "evi_swin11_annual"
names(annual_rds_list[[2]])<- "ndvi_swin11_annual"

##Analyses
cropmask_function <- function (trajectory_raster){
  x_crop <- terra::crop(trajectory_raster, cvnp_shp)
  x_mask <- terra::mask (x_crop, cvnp_shp)
  x_mask
}
output_file_path <- here("Outputs", "TrendsResults", "cvnp_results")

Sys.time(); cvnp_annual_evi <- cropmask_function(annual_rds_list[[1]]); Sys.time()
#writeRaster(cvnp_annual_evi, here(output_file_path,"cvnp_evi_annual.tif"))
Sys.time(); cvnp_annual_ndvi <- cropmask_function(annual_rds_list[[2]]); Sys.time()
#writeRaster(cvnp_annual_ndvi, here(output_file_path,"cvnp_ndvi_annual.tif"))
Sys.time(); cvnp_monthly_evi <- cropmask_function(monthly_rds_list[[1]]); Sys.time()
#writeRaster(cvnp_monthly_evi, here(output_file_path,"cvnp_evi_monthly.tif"))
Sys.time(); cvnp_monthly_ndvi <- cropmask_function(monthly_rds_list[[2]]); Sys.time()
#writeRaster(cvnp_monthly_ndvi, here(output_file_path,"cvnp_ndvi_monthly.tif"))

remove(annual_file_list, annual_input_file_path, monthly_file_list, monthly_input_file_path)
remove(annual_rds_list, monthly_rds_list, cropmask_function)
```


#Step 2- mapping trajectory results in CVNP
```{r}
mapping_function <- function (trendresults_raster, title){
  x_map<- tm_shape(cvnp_shp)+ tm_borders()+
    tm_shape (trendresults_raster) + 
    tm_raster(col.scale = tm_scale_categorical(n=6, values = c( "#117733", "lightgrey", "#882255", "#88CCEE", "#332288", "#DDCC77" )), col.legend = tm_legend_hide()) +
    tm_add_legend( fill = c("#117733", "lightgrey", "#882255", "#88CCEE", "#332288", "#DDCC77"),
      labels = c("Linear increase",
                           "No trend",
                           "Step decrease",
                           "Step increase",
                           "Quad increase (accelerated)",
                           "Quad increase (decelerated)")) + 
    tm_layout(legend.text.size = 2) + tm_title(title)
  x_map
}

Sys.time();cvnp_monthly_ndvi_results_map <- mapping_function (cvnp_monthly_ndvi, "NDVI (monthly)"); Sys.time()
Sys.time();cvnp_monthly_evi_results_map <- mapping_function (cvnp_monthly_evi, "EVI (monthly)"); Sys.time()
Sys.time();cvnp_annual_ndvi_results_map <- mapping_function (cvnp_annual_ndvi, "NDVI (annual)"); Sys.time()
Sys.time();cvnp_annual_evi_results_map <- mapping_function (cvnp_annual_evi, "EVI (annual)"); Sys.time()
cvnp_trendresult_map<- tmap_arrange(cvnp_monthly_ndvi_results_map, 
                                    cvnp_annual_ndvi_results_map, 
                                    cvnp_monthly_evi_results_map, 
                                    cvnp_annual_evi_results_map,ncol =2, nrow= 2)
output_file_path <- here("Outputs", "TrendsResults", "cvnp_results")
tmap_save(cvnp_trendresult_map, paste0(output_file_path, "/", "cvnp_map1km_trendresults_swin11_monthlyannual.png"),
        height = 80, width = 80, units = "cm", dpi=700)


remove(cvnp_annual_evi_results_map, cvnp_annual_ndvi_results_map, cvnp_monthly_evi_results_map, cvnp_monthly_ndvi_results_map, cvnp_trendresult_map, mapping_function)
```


#Step 3- barplot results of area of different trajectories
```{r}
df_prep_function<- function (trendresults_raster, index_name){
  x_df<- terra::as.data.frame(trendresults_raster, cell = TRUE, xy= TRUE)
  x_df<- x_df %>% mutate(index = index_name)
  names(x_df)[4]<- "value"
  percentage_df <- x_df %>% group_by(value, index) %>% summarise(count= n()) %>% mutate(percentage= (count/nrow(x_df))*100) 
  percentage_df
}

Sys.time(); cvnp_monthly_ndvi_trend_results <- df_prep_function(cvnp_monthly_ndvi, "NDVI (monthly)"); Sys.time()
Sys.time(); cvnp_monthly_evi_trend_results <- df_prep_function(cvnp_monthly_evi, "EVI(monthly)"); Sys.time()
Sys.time(); cvnp_annual_ndvi_trend_results <- df_prep_function(cvnp_annual_ndvi, "NDVI (annual)"); Sys.time()
Sys.time(); cvnp_annual_evi_trend_results <- df_prep_function(cvnp_annual_evi, "EVI(annual)"); Sys.time()

plot_df <- bind_rows(cvnp_monthly_ndvi_trend_results, cvnp_monthly_evi_trend_results,
                     cvnp_annual_ndvi_trend_results, cvnp_annual_evi_trend_results)

plot_df <- plot_df %>% mutate(value = case_when(value==1~ "Linear decrease",
                                                value==2~ "Linear increase",
                                                value==4~ "Step decrease",
                                                value==5~"Step increase",
                                                value==6~"Quadratic decrease (accelerated)",
                                                value==7~"Quadratic decrease (decelerated)",
                                                value==8~"Quadratic increase (accelerated)",
                                                value==9~"Quadratic increase (decelerated)",
                                                TRUE~"No trend"))

Set1palette<- c("#117733", "lightgrey", "#332288", "#DDCC77", "#882255","#88CCEE")
cvnp_trendresults_barplot<- plot_df %>% 
  ggplot(aes(value, percentage, fill = value, group=index, alpha=as.factor(index))) +
  geom_col(position = position_dodge(), color = "black") +
  theme_classic() +
  xlab("Trajectory Shape")+ ylab ("% of native savannas")+
  theme(strip.placement  = "outside",
        panel.spacing    = unit(0, "points"),
        strip.background = element_blank(),
        strip.text       = element_text(size = 12, angle = 90),
        axis.text.x = element_text(size = 12, angle = 90)) +
  scale_fill_manual(values = Set1palette) + theme(legend.title=element_blank()) +
  xlab("Model Type")
ggsave(here(output_file_path, paste0("cvnp_barplot_monthlyswin11_trendresults.png")),
       cvnp_trendresults_barplot,dpi = 700, height = 20, width=20, units = "cm")

remove(cvnp_annual_evi_trend_results, cvnp_annual_ndvi_trend_results, cvnp_monthly_evi_trend_results, cvnp_monthly_ndvi_trend_results, Set1palette, cvnp_trendresults_barplot, df_prep_function)
remove(cvnp_annual_evi, cvnp_annual_ndvi, cvnp_monthly_evi, cvnp_monthly_ndvi, plot_df)

```
