==> Scripts/Final_Scripts/cvnp_analyses-RETAIN BUT NOT IN THIS REPOSITORY/4_trajectory_heterogeniety_relationship.Rmd <==
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r echo=FALSE, include=FALSE}
library(RColorBrewer)
library(lattice)
library(tidyverse)
library(here)

library(ggplot2)
library(ggpubr)
library(sf)
library(terra)

library(tmap)
library(tmaptools)
library(viridis)

terraOptions(memfrac=0.8, tempdir = here("Scratch"), progress=10)
```
##Introduction
In this script, I complete analyses to explore the relationship between the trajectories and pixel heterogeniety based on vegetation physiognomies from Lewis et al., 2020
(1) Data input of all trajectory results
(2) Lewis et al data input and processing to calculate pixel heterogeniety
(3) EDA relationships

#Step 1- data input of trajectory results
```{r}
cvnp_shp<- st_read(here("Data", "Admin", "WDPA_WDOECM_Jan2024_Public_59_shp","WDPA_WDOECM_Jan2024_Public_59_shp_0", "WDPA_WDOECM_Jan2024_Public_59_shp-polygons.shp")) 

input_file_path <- here("Outputs", "TrendsResults", "cvnp_results")

cvnp_annual_evi<- rast(here(input_file_path,"cvnp_evi_annual.tif"))
cvnp_annual_ndvi<- rast(here(input_file_path,"cvnp_ndvi_annual.tif"))
cvnp_monthly_evi<- rast(here(input_file_path,"cvnp_evi_monthly.tif"))
cvnp_monthly_ndvi<- rast(here(input_file_path,"cvnp_ndvi_monthly.tif"))

```

#Step 2 - veg physiognomies input and processing to calcualte pixel heterogeniety (only considering native physiognomies)
```{r}
veg_phy<- rast(here("Data", "Lewisetal2022_ChapadaNP_PhysiognomyMaps", "PNCV_vegMap_woText.tif"))
#0= campo umido, 1=campo sujo, 3=campu rupestre, 5=cerrado sensu stricto, 6=cerrado rupestre, 8=verada
#9=cerradao, 10=mata de galeria, 11-pasture, 12- water, 13- plantation, 14- agriculture, 15- nonvegetated

crop_veg_phy<- terra::crop(veg_phy, vect(cvnp_shp))
mask_veg_phy<- terra::mask(crop_veg_phy, vect(cvnp_shp))
remove(crop_veg_phy, veg_phy)

m_nativephysiognomies_only<- rbind(c(0,0), c(1,1), c(3,3), c(5,5), c(6,6), c(8,8),
                                   c(9,9), c(10,10), c(11,NA), c(12,NA), c(13,NA), c(14,NA), c(15, NA)) # Excluding anthropic physiognomies and calculating richness of native physiognomies
reclass_veg_phy <- terra::classify(mask_veg_phy, m_nativephysiognomies_only)
remove(m_nativephysiognomies_only)

coarsening_function <- function (classified_raster){
  x_1km<- terra::project(
    terra::segregate(classified_raster), cvnp_annual_evi,
    method = "average", res = res(cvnp_annual_evi)[1])
 
  x_1km_mask <- terra::mask(x_1km, cvnp_shp)
  x_1km_mask
}

coarse_reclass_veg_phy <- coarsening_function(reclass_veg_phy)
remove(coarsening_function)

library(vegan) ##Shannon's diversity index
shannon_diversity<- function (reclass_coarse_raster) {
    x_df<- as.data.frame(reclass_coarse_raster, cells=TRUE, xy=TRUE)
    x_diversity<- vegan::diversity(x_df %>% dplyr::select(-c(cell,x,y)), index="shannon")
    x_diversity<- as.data.frame(x_diversity)
    x_df<- bind_cols(x_df, x_diversity)
    trial_vector<- terra::vect(x_df %>% dplyr::select(c("cell", "x", "y", "x_diversity")), geom=c("x", "y"), crs="epsg:4326")
    trial_raster<- terra::rasterize(trial_vector, reclass_coarse_raster, "x_diversity", fun="max")
    trial_raster
}
Sys.time(); heterogen <- shannon_diversity (coarse_reclass_veg_phy);Sys.time()
remove(shannon_diversity)

x_map<- tm_shape(cvnp_shp)+ tm_borders()+
    tm_shape (heterogen) + 
    tm_raster(col.scale = tm_scale_intervals(values = "Oranges"),
              col.legend = tm_legend(title = "Native veg heterogeniety")) +
    tm_layout(legend.text.size = 1, title.size = 4) 
tmap_save(x_map, here("Outputs", "OtherVariables", "Formations_Heterogeniety", "cvnp_pixelheterogeniety.png"),
        height = 50, width = 50, units = "cm", dpi=700)


```

#Step 3a- heterogeneity spread in pixels classified as each of the trajectories
```{r}
data_stack <- c(cvnp_annual_evi, cvnp_annual_ndvi,cvnp_monthly_evi,cvnp_monthly_ndvi, heterogen)
data_df <- terra::as.data.frame(data_stack)
data_df_noNA <- data_df %>% drop_na()
names(data_df_noNA) <- c("annual_EVI", "annual_NDVI", "monthly_EVI", "monthly_NDVI", "heterogeniety")

heter_trajtype_boxplot_function <- function (traj_column, plot_title){
  df_filter <- data_df_noNA  %>% dplyr::select(heterogeniety, traj_column)
  df_filter <- df_filter %>% mutate(TrajType = case_when(df_filter[[traj_column]] == 2~" Linear Increase",
                                                                         df_filter[[traj_column]] == 3 ~"No trend",
                                                                         df_filter[[traj_column]] == 4 ~"Step Decrease",
                                                                         df_filter[[traj_column]] == 5 ~ "Step Increase",
                                                                         df_filter[[traj_column]] == 8 ~ "Quadratic Increase (acc)",
                                                                         df_filter[[traj_column]] == 9 ~ "Quadratic Increase (dec)"))

  plot<- ggplot(df_filter, aes(x = reorder(TrajType, heterogeniety, FUN = median), y = heterogeniety, fill = TrajType)) + 
    geom_boxplot() + ylab("native vegetation heterogeniety") + xlab("Trajectory types") +
    labs(title = plot_title) +
    theme_classic(base_size = 16) +
     theme(axis.text.x =element_blank(),
        axis.ticks.x =element_blank()) +
    scale_fill_manual(values=c("#117733", "lightgrey", "#332288", "#DDCC77", "#882255", "#88CCEE"))
  plot 
}

box_ndvi_monthly <- heter_trajtype_boxplot_function("monthly_NDVI", "NDVI(monthly)")
box_evi_monthly <- heter_trajtype_boxplot_function("monthly_EVI", "EVI(monthly)")
box_ndvi_annual <- heter_trajtype_boxplot_function("annual_NDVI", "NDVI(annual)")
box_evi_annual <- heter_trajtype_boxplot_function("annual_EVI", "EVI(annual)")

boxplot_plots <- ggarrange(box_ndvi_monthly, box_evi_monthly, box_ndvi_annual, box_evi_annual, ncol =2, nrow=2 )
output_file_path <- here("Outputs", "TrendsResults", "cvnp_results")
ggsave(paste0(output_file_path, "//", "boxplot_trajectory_heterogeniety.png"),
       boxplot_plots ,dpi = 700, height = 80, width = 80, units = "cm")

remove(box_ndvi_annual, box_evi_annual, box_evi_monthly, box_ndvi_monthly, boxplot_plots, heter_trajtype_boxplot_function)
```

#Step 3b- proportion of pixels in particular bins of hetergeniety and the trajectory that the pixels belong to
```{r}
data_df <- data_df %>% mutate(bins = cut(max, breaks = quantile(max, na.rm= TRUE), right = FALSE, labels = FALSE, ordered_result = TRUE))
data_df <- data_df %>% mutate(bins = ifelse(is.na(bins), 4, bins))
#data_df <- data_df %>% mutate(bins = ifelse(is.na(bins), "No heterogeniety", bins))
number_pixels_bin <- data_df %>% group_by(bins) %>% summarise(count=n())

pivot_data_df <- data_df %>% pivot_longer(1:4)
heter_trajectory_df <- pivot_data_df %>%
  group_by(name, bins, value) %>%
  summarise(total_pixels = n())

heter_trajectory_df  <- inner_join(heter_trajectory_df , number_pixels_bin)
heter_trajectory_df <- heter_trajectory_df %>% mutate(Proportion= (total_pixels/count)*100) #nrow 
heter_trajectory_df <- heter_trajectory_df %>% filter(!is.na(value))
heter_trajectory_df <- heter_trajectory_df %>% mutate(TrajType = case_when(value == 2~" Linear Increase",
                                                                         value == 3 ~"No trend",
                                                                         value == 4 ~"Step Decrease",
                                                                         value == 5 ~ "Step Increase",
                                                                         value == 8 ~ "Quadratic Increase (acc)",
                                                                         value == 9 ~ "Quadratic Increase (dec)"))
heter_trajectory_df <- heter_trajectory_df %>% mutate(bins = case_when (bins == 1~"1st quantile",
                                                                        bins == 2 ~ "2nd quantile",
                                                                        bins == 3 ~ "3rd quantile",
                                                                        bins == 4 ~ "4th quantile",
                                                                        TRUE ~"No heterogeniety"))
heter_trajectory_df <- heter_trajectory_df %>% mutate(name = case_when (name == "evi_swin11_annual" ~"EVI(annual)",
                                                                        name == "evi_swin11_monthly" ~"EVI(monthly)",
                                                                        name == "ndvi_swin11_annual" ~"NDVI(annual)",
                                                                        TRUE ~ "NDVI(monthly)"))

  

plot <- ggplot(data = heter_trajectory_df, aes(x = bins, y= Proportion, fill = TrajType))+
    geom_bar(stat = "identity", position = position_dodge()) + 
  theme_classic(base_size = 16) +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1)) +
  xlab("Heterogeniety")+
  scale_fill_manual(values=c("#117733", "lightgrey", "#332288", "#DDCC77", "#882255", "#88CCEE"))+
  facet_grid(~name)
ggsave(paste0(output_file_path, "//", "relationship_trajectory_heterogeniety.png"),
       plot ,dpi = 700, height = 80, width = 80, units = "cm")




```


#Step 4- statistical analyses of relationship between heterogeniety and trajectory types (categorical variable) using anova
```{r}
ndvimonthlytraj_hetero_aov<- lm(max ~ as.factor(ndvi_swin11_monthly),
                                data_df %>% filter(ndvi_swin11_monthly == 5 | ndvi_swin11_monthly == 3)) 
summary(ndvimonthlytraj_hetero_aov)
```

No difference in group means of heterogeniety in step increase and no trend trajectories.
